'have' commit 3/1/14
Created functionality so that if peer A receives a ‘have’ message from
peer B for a piece it does not have (using method ‘hasPiece() to check.
Method does not exist yet, but should be in another class), A sends an
‘interested’ message to C. Else, A sends a ‘not interested’ message.
This method also adds information to the log file.
The method should also allow the receiving peer to update bitfields to
all neighbors. This has not been implemented yet as there is not a
class for bitfield yet.